+++ packages/adhere-mobile-ui-anthoc/src/form/Form.ts
@@ -5,6 +5,7 @@ import Validator from '@baifendian/adhere-util-validator';
 
 import type { FormHOCComponent, FormValidatorRulesType } from '../types';
 import { createFactory } from '../util';
+import NestingFormItem from './NestingFormItem';
 
 const FormHOC: FormHOCComponent = createFactory<FormProps>(Form, {});
 
@@ -84,6 +85,7 @@ let ValidatorRules: FormValidatorRulesType;
 })();
 
 FormHOC.ValidatorRules = ValidatorRules;
+FormHOC.NestingFormItem = NestingFormItem;
 FormHOC.displayName = 'Form';
 
 export default FormHOC;