+++ packages/adhere-mobile-ui-anthoc/src/form/NestingFormItem.tsx
@@ -54,6 +54,7 @@ const InternalNestingFormItem = memo<
 
       return (
         <Form
+          // @ts-ignore
           ref={ref}
           className={className ?? ''}
           style={style ?? {}}